# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Awaaz Setu API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ReportInput {
  category: string;
  district: string;
  statement: string;
  evidenceFiles?: string[];
}

export interface ReportSubmitResult {
  token: string;
  message: string;
}

export interface TrackingEvent {
  status: string;
  timestamp: string;
  description: string;
  /** @nullable */
  severity?: string | null;
}

export interface ReportTrackResult {
  token: string;
  category: string;
  district: string;
  submittedAt: string;
  currentStatus?: string;
  events: TrackingEvent[];
}

export interface AdminReport {
  id: string;
  token: string;
  category: string;
  district: string;
  submittedAt: string;
  severity: string;
  status: string;
  /** @nullable */
  assignedTo?: string | null;
}

export interface EvidenceFile {
  name: string;
  type: string;
  /** @nullable */
  size?: string | null;
}

export interface AdminReportDetail {
  id: string;
  token: string;
  category: string;
  district: string;
  submittedAt: string;
  severity: string;
  status: string;
  statement: string;
  /** @nullable */
  assignedTo?: string | null;
  evidenceFiles?: EvidenceFile[];
  timeline: TrackingEvent[];
}

export interface RouteAction {
  action: string;
  /** @nullable */
  note?: string | null;
}

export interface CategoryStat {
  category: string;
  count: number;
  percentage: number;
}

export interface AdminStats {
  totalReports: number;
  activeCritical: number;
  resolved: number;
  pendingTriage: number;
  categoryBreakdown: CategoryStat[];
  severityBreakdown: CategoryStat[];
  recentActivity: AdminReport[];
}

export interface StateStat {
  state: string;
  count: number;
  topCategory: string;
}

export interface AwarenessStats {
  totalIncidents: number;
  statesReporting: number;
  topCategory: string;
  byState: StateStat[];
  byCategory: CategoryStat[];
  disclaimer: string;
}

export type TrackReportParams = {
token: string;
};