# lib/api-zod/src/generated/types/adminReport.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Awaaz Setu API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminReport {
  id: string;
  token: string;
  category: string;
  district: string;
  submittedAt: string;
  severity: string;
  status: string;
  /** @nullable */
  assignedTo?: string | null;
}