# lib/api-zod/src/generated/types/adminReportDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Awaaz Setu API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EvidenceFile } from './evidenceFile';
import type { TrackingEvent } from './trackingEvent';

export interface AdminReportDetail {
  id: string;
  token: string;
  category: string;
  district: string;
  submittedAt: string;
  severity: string;
  status: string;
  statement: string;
  /** @nullable */
  assignedTo?: string | null;
  evidenceFiles?: EvidenceFile[];
  timeline: TrackingEvent[];
}