# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Awaaz Setu API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminReport } from './adminReport';
import type { CategoryStat } from './categoryStat';

export interface AdminStats {
  totalReports: number;
  activeCritical: number;
  resolved: number;
  pendingTriage: number;
  categoryBreakdown: CategoryStat[];
  severityBreakdown: CategoryStat[];
  recentActivity: AdminReport[];
}