# lib/api-zod/src/generated/types/awarenessStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Awaaz Setu API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryStat } from './categoryStat';
import type { StateStat } from './stateStat';

export interface AwarenessStats {
  totalIncidents: number;
  statesReporting: number;
  topCategory: string;
  byState: StateStat[];
  byCategory: CategoryStat[];
  disclaimer: string;
}