# lib/api-zod/src/generated/types/evidenceFile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Awaaz Setu API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EvidenceFile {
  name: string;
  type: string;
  /** @nullable */
  size?: string | null;
}