# lib/api-zod/src/generated/types/reportSubmitResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Awaaz Setu API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReportSubmitResult {
  token: string;
  message: string;
}