# lib/api-zod/src/generated/types/reportTrackResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Awaaz Setu API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TrackingEvent } from './trackingEvent';

export interface ReportTrackResult {
  token: string;
  category: string;
  district: string;
  submittedAt: string;
  currentStatus?: string;
  events: TrackingEvent[];
}