# lib/api-zod/src/generated/types/stateStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Awaaz Setu API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StateStat {
  state: string;
  count: number;
  topCategory: string;
}